requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
decorator==4.4.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
geocoder==1.38.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gTTS==2.2.1
httplib2==0.18.1
idna==2.10
numpy==1.19.5
opencv-python==4.5.1.48
pafy==0.5.5
Pillow==8.1.0
playsound==1.2.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
python-vlc==3.0.11115
pyttsx3==2.90
pytz==2020.5
ratelim==0.1.6
requests==2.25.1
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.1
uritemplate==3.0.1
urllib3==1.26.2
vext==0.7.4
vext.gi==0.7.4
youtube-dl==2021.1.8
